+++ source/Examples/Resources/SqlSecureConnection/4-SecureConnectionUsingSYSTEMAccount.ps1
@@ -10,7 +10,7 @@ Configuration Example
 
     node localhost
     {
-        SqlServerSecureConnection 'SecureConnectionUsingSYSTEMAccount'
+        SqlSecureConnection 'SecureConnectionUsingSYSTEMAccount'
         {
             InstanceName    = 'MSSQLSERVER'
             Thumbprint      = 'fb0b82c94b80da26cf0b86f10ec0c50ae7864a2c'